+++ lib/Firebug/index.jsx
@@ -1,4 +1,4 @@
-import Button from '@mui/material/Button'
+import { Button } from '@mui/material'
 import Logo from './firebug.svg'
 
 export default function Firebug() {
