+++ lib/SomDatePicker/index.jsx
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
-import { ArrowBackIosOutlined, ArrowForwardIosOutlined } from '@mui/icons-material'
-import Box from '@mui/material/Box'
-import Button from '@mui/material/Button'
+import ArrowBackIosOutlined from '@mui/icons-material/ArrowBackIosOutlined'
+import ArrowForwardIosOutlined from '@mui/icons-material/ArrowForwardIosOutlined'
+import { Box, Button } from '@mui/material'
 import { DatePicker } from '@mui/x-date-pickers/DatePicker'
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs'
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider'
