+++ lib/TableEditor/index.jsx
@@ -2,16 +2,7 @@
 
 import React from 'react'
 import PropTypes from 'prop-types'
-import Box from '@mui/material/Box'
-import Table from '@mui/material/Table'
-import TableBody from '@mui/material/TableBody'
-import TableCell from '@mui/material/TableCell'
-import TableContainer from '@mui/material/TableContainer'
-import TablePagination from '@mui/material/TablePagination'
-import TableRow from '@mui/material/TableRow'
-import Paper from '@mui/material/Paper'
-import Checkbox from '@mui/material/Checkbox'
-import Button from '@mui/material/Button'
+import { Box, Table, TableBody, TableCell, TableContainer, TablePagination, TableRow, Paper, Button, Checkbox } from '@mui/material'
 import { useTranslation } from '../i18n'
 import { ActionsType } from './proptypes'
 import ActionButtons from './ActionButtons'
